Fix case where second command doesn't take stdin
e.g. `ss | ls`

If the second command has exited, it's stdin has also exited. So we
can no longer write to it's stdin.
We fixed this by introducing the try_wait check, but when the second
command exits between the try_wait check and writing to stdin, we may
still get an error.